# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * 8848LABS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  description?: string;
  shortDescription?: string;
  category: string;
  images: string[];
  materials: string[];
  colors?: string[];
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  printTime?: string | null;
  inStock: boolean;
  /** @nullable */
  stock?: number | null;
  featured?: boolean;
  tags?: string[];
  customizable?: boolean;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  reviewCount?: number | null;
  createdAt?: string;
}

export interface ProductList {
  products: Product[];
  total: number;
}

export interface Collection {
  id: number;
  name: string;
  slug: string;
  description: string;
  image: string;
  productCount?: number;
  featured?: boolean;
}

export interface CollectionDetail {
  id: number;
  name: string;
  slug: string;
  description: string;
  image: string;
  products: Product[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image?: string | null;
  productCount: number;
}

export interface Project {
  id: number;
  title: string;
  slug: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  images: string[];
  tags: string[];
  /** @nullable */
  material?: string | null;
  /** @nullable */
  client?: string | null;
  featured?: boolean;
  /** @nullable */
  number?: string | null;
  createdAt?: string;
}

export interface JournalPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  /** @nullable */
  content?: string | null;
  coverImage: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  readTime?: string | null;
  publishedAt: string;
}

export type CustomOrderInputDimensionUnit = typeof CustomOrderInputDimensionUnit[keyof typeof CustomOrderInputDimensionUnit];


export const CustomOrderInputDimensionUnit = {
  mm: 'mm',
  cm: 'cm',
  inches: 'inches',
} as const;

export interface CustomOrderInput {
  projectName: string;
  description: string;
  /** @nullable */
  intendedUse?: string | null;
  /** @nullable */
  lengthMm?: number | null;
  /** @nullable */
  widthMm?: number | null;
  /** @nullable */
  heightMm?: number | null;
  dimensionUnit?: CustomOrderInputDimensionUnit;
  quantity?: number;
  preferredMaterial?: string;
  /** @nullable */
  preferredColor?: string | null;
  desiredFinish?: string;
  /** @nullable */
  budgetRange?: string | null;
  /** @nullable */
  desiredDeliveryDate?: string | null;
  /** @nullable */
  additionalNotes?: string | null;
  fullName: string;
  email: string;
  /** @nullable */
  phone: string | null;
  preferredContact?: string;
  fileUrls?: string[];
}

export interface CustomOrder {
  id: number;
  projectName: string;
  status: string;
  createdAt: string;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  personalization?: string | null;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  productImage: string;
  price: number;
  quantity: number;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  personalization?: string | null;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  itemCount: number;
}

export interface Testimonial {
  id: number;
  name: string;
  initials?: string;
  /** @nullable */
  avatar?: string | null;
  rating: number;
  text: string;
  /** @nullable */
  productPurchased?: string | null;
  /** @nullable */
  location?: string | null;
}

export interface Stats {
  customersServed: number;
  projectsCompleted: number;
  customDesigns: number;
  printingHours: number;
  satisfactionPercent: number;
  avgProductionDays: number;
}

export type ListProductsParams = {
category?: string;
material?: string;
color?: string;
minPrice?: number;
maxPrice?: number;
sort?: string;
search?: string;
inStock?: boolean;
limit?: number;
offset?: number;
};