# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 8848LABS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: material
          in: query
          schema: { type: string }
        - name: color
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: sort
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: inStock
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer }
        - name: offset
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
...
  /journal/{slug}:
    get:
      operationId: getJournalPost
      tags: [journal]
      summary: Get a single journal post
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Journal post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalPost"
        "404":
          description: Not found

  # ── Custom Orders ─────────────────────────────────────────────────────────
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ── Testimonials ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, price, category, images, inStock, materials]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        price: { type: number }
        compareAtPrice: { type: ["number", "null"] }
        description: { type: string }
        shortDescription: { type: string }
        category: { type: string }
        images: { type: array, items: { type: string } }
        materials: { type: array, items: { type: string } }
        colors: { type: array, items: { type: string } }
        dimensions: { type: ["string", "null"] }
        weight: { type: ["string", "null"] }
        printTime: { type: ["string", "null"] }
        inStock: { type: boolean }
        stock: { type: ["integer", "null"] }
        featured: { type: boolean }
        tags: { type: array, items: { type: string } }
        customizable: { type: boolean }
        rating: { type: ["number", "null"] }
        reviewCount: { type: ["integer", "null"] }
        createdAt: { type: string }
    ProductList:
      type: object
      required: [products, total]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer }

    # ── Collection ────────────────────────────────────────────────────────────
...
    JournalPost:
      type: object
      required: [id, title, slug, excerpt, coverImage, publishedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: ["string", "null"] }
        coverImage: { type: string }
        category: { type: ["string", "null"] }
        readTime: { type: ["string", "null"] }
        publishedAt: { type: string }

    # ── Custom Order ──────────────────────────────────────────────────────────
...
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
        color: { type: ["string", "null"] }
        material: { type: ["string", "null"] }
        personalization: { type: ["string", "null"] }
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer }
    CartItem:
      type: object
      required: [id, productId, productName, productImage, price, quantity]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        productImage: { type: string }
        price: { type: number }
        quantity: { type: integer }
        color: { type: ["string", "null"] }
        material: { type: ["string", "null"] }
        personalization: { type: ["string", "null"] }
    Cart:
      type: object
      required: [items, subtotal, itemCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal: { type: number }
        itemCount: { type: integer }

    # ── Testimonial ───────────────────────────────────────────────────────────
...