# lib/api-zod/src/generated/types/cartItemInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * 8848LABS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemInput {
  productId: number;
  quantity: number;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  material?: string | null;
  /** @nullable */
  personalization?: string | null;
}